+++ tinygrad/tensor.py
@@ -1,5 +1,6 @@
 # inspired by https://github.com/karpathy/micrograd/blob/master/micrograd/engine.py
 from __future__ import annotations
+import os
 import time
 from functools import partialmethod, reduce
 from itertools import accumulate, filterfalse
@@ -219,7 +220,7 @@ class Tensor:
 
     # fill in the first grad with one. don't use Tensor.ones because we don't need contiguous
     # this is "implicit gradient creation"
-    self.grad = Tensor(1, device=self.device, requires_grad=False)
+    self.grad = Tensor(1, device=self.device, requires_grad=False, dtype=self.dtype)
 
     for t0 in reversed(self.deepwalk()):
       if not t0.requires_grad:
